+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Fixed
 
+- type mismatch errors inside the `copy` function are not shown
+
 ### Removed
 
 ### Updated